Implementation of Pathfind
Pathfinding check has been added. Likely to be the final implementation of the targetting.

Pathfinding is strictly checking for path within 1 block during damage events. If there exists no path between the player and entity, it is assumed the player cannot be attacked.

Created registerCommand() to get rid of the annoying IntelliJ warning wanting me to do Objects.requireNonNull() and no I am totally not angry.